move x86_64 ones to x86_64
