+++ crates/core_arch/avx512bw.md
@@ -1,34 +1,34 @@
 <summary>["AVX512BW"]</summary><p>
 
   * [x] [`_mm512_loadu_epi16`]
-  * [_] [`_mm512_mask_loadu_epi16`]
-  * [_] [`_mm512_maskz_loadu_epi16`]
+  * [_] [`_mm512_mask_loadu_epi16`] //need i1
+  * [_] [`_mm512_maskz_loadu_epi16`] //need i1
   * [x] [`_mm_loadu_epi16`]
-  * [_] [`_mm_mask_loadu_epi16`]
-  * [_] [`_mm_maskz_loadu_epi16`]
+  * [_] [`_mm_mask_loadu_epi16`] //need i1
+  * [_] [`_mm_maskz_loadu_epi16`] //need i1
   * [x] [`_mm256_loadu_epi16`]
-  * [_] [`_mm256_mask_loadu_epi16`]
-  * [_] [`_mm256_maskz_loadu_epi16`]
+  * [_] [`_mm256_mask_loadu_epi16`] //need i1
+  * [_] [`_mm256_maskz_loadu_epi16`] //need i1
   * [x] [`_mm512_loadu_epi8`]
-  * [_] [`_mm512_mask_loadu_epi8`]
-  * [_] [`_mm512_maskz_loadu_epi8`]
+  * [_] [`_mm512_mask_loadu_epi8`] //need i1
+  * [_] [`_mm512_maskz_loadu_epi8`] //need i1
   * [x] [`_mm_loadu_epi8`]
-  * [_] [`_mm_mask_loadu_epi8`]
-  * [_] [`_mm_maskz_loadu_epi8`]
+  * [_] [`_mm_mask_loadu_epi8`] //need i1
+  * [_] [`_mm_maskz_loadu_epi8`] //need i1
   * [x] [`_mm256_loadu_epi8`]
-  * [_] [`_mm256_mask_loadu_epi8`]
-  * [_] [`_mm256_maskz_loadu_epi8`]
+  * [_] [`_mm256_mask_loadu_epi8`] //need i1
+  * [_] [`_mm256_maskz_loadu_epi8`] //need i1
   * [_] [`_mm512_mask_storeu_epi16`]
   * [x] [`_mm512_storeu_epi16`]
-  * [_] [`_mm_mask_storeu_epi16`]
+  * [_] [`_mm_mask_storeu_epi16`] //need i1
   * [x] [`_mm_storeu_epi16`]
-  * [_] [`_mm256_mask_storeu_epi16`]
+  * [_] [`_mm256_mask_storeu_epi16`] //need i1
   * [x] [`_mm256_storeu_epi16`]
-  * [_] [`_mm512_mask_storeu_epi8`]
+  * [_] [`_mm512_mask_storeu_epi8`] //need i1
   * [x] [`_mm512_storeu_epi8`]
-  * [_] [`_mm_mask_storeu_epi8`]
+  * [_] [`_mm_mask_storeu_epi8`] //need i1
   * [x] [`_mm_storeu_epi8`]
-  * [_] [`_mm256_mask_storeu_epi8`]
+  * [_] [`_mm256_mask_storeu_epi8`] //need i1
   * [x] [`_mm256_storeu_epi8`]
   * [x] [`_mm512_abs_epi16`]
   * [x] [`_mm512_mask_abs_epi16`]
